requirements.txt
asyncpg
better-profanity
discord.py
jishaku
python-dotenv